Add files via upload
In South Africa, the department of transport issues different driving licenses depending on the 
category (or categories) of vehicle that the recipient wishes to drive (or learn to drive).
How is a vehicle classified? It depends on the Gross Vehicle Mass (GVM), and on whether the vehicle 
is ‘articulated’, or whether a trailer is being towed. 
The GVM dictates the base class to which the vehicle belongs: B, C1, or C.
GVM≤3500 Class B
3500<GVM≤16000 Class C1
16000<GVM Class C
If, however, the vehicle is articulated, or is towing a trailer with a GVM greater than 750kg, then it 
belongs in an extended, ‘E’, version of the class: EB, EC1, or EC. (According to Wikipedia, an 
articulated vehicle is one “which has a permanent or semi-permanent pivoting joint in its 
construction”.)
4

The classification scheme does not allow for an articulated vehicle towing a trailer. It’s an ‘eitheror’ situation.
Write a Python program called ‘vehicleclassifier.py’ that can calculate the class to which a 
vehicle belongs.